refactor runtime of deltann
